HT_24/requirements.txt
celery==5.3.6
django==5.0
django-celery-results==2.5.1
django-environ==0.11.2
djangorestframework==3.14.0
factory_boy==3.3.0
redis==5.0.1
requests==2.31.0